>Added proper occurrence handling. >Fixed bugs related to adding and removing patterns/occurrences. >Reworked menu items. >Added functionality for occurrence showing/hiding, highlighting, and collapse/expand. >Added functionality for deleting patterns/occurrences. >Added zoom capability. >Added a timeline bar at the top. >Added the ability to change snap to grid. >Implemented functionality to find duplicates of a certain occurrence. >Changed icon at the welcome screen. >Improved visual clarity.

+++ annotation-tool/annotation-tool/classes/parsers/NoteParser.cs
@@ -31,7 +31,7 @@ namespace AnnotationTool
                 {
                     ChannelEvent tempNoteEvent = (ChannelEvent)tempEvent;
                     Note currentNote = EventToNote(tempNoteEvent);
-                    currentNote.SetTime(timeElapsed);
+                    currentNote.SetTime((double)timeElapsed);
                     
                     // If there is already an active note in this pitch, add the new note to the existing stack.
                     if (activeNotes.ContainsKey(currentNote.GetPitch()))
@@ -54,7 +54,7 @@ namespace AnnotationTool
                     if (activeNotes.ContainsKey(currentNote.GetPitch()))
                     {
                         Note tempNote = activeNotes[currentNote.GetPitch()].Pop();
-                        tempNote.SetDuration(timeElapsed - tempNote.GetTime());
+                        tempNote.SetDuration((double)timeElapsed - tempNote.GetTime());
                         notes.Add(tempNote);
 
                         if (activeNotes[currentNote.GetPitch()].Count == 0) // If the stack of notes is empty, remove the entry from the dictionary of active notes.
